+++ src/Address/MuxedAddress.h
@@ -11,7 +11,7 @@ namespace ArduinoIO
 	{
 		class MuxedAddress : public AddressBase
 		{
-			public:MuxedAddress(MultiplexerBase* mux, byte muxPort);
+			public:MuxedAddress(MultiplexerBase* mux, int muxPort);
 			public:virtual void PrepareForReading();//chiama mux.set
 			protected: MultiplexerBase* _mux;
 			protected: byte _port;
